refactor(xtask): share build helpers between WASM and WASI pipelines
build.rs and build_wasi.rs had copy-pasted, silently-drifted helpers. Hoist the
shared logic into util.rs:

- find_occt_lib_dir: the two copies had divergent candidate lists/order
  (build.rs tried lin32/lin/lib; build_wasi tried lib/lin32/wasm32), so on a
  machine with a stray top-level occt/build/lib they could resolve to different
  directories. Unified to one ordering that tries the clang paths before bare
  lib. In the CI/builder layout (libs in lin32/clang/lib) both pipelines resolve
  to the same directory as before — the WASI stale-check verifies byte-identical
  output.
- find_wasm_opt + home_dir: emsdk-first wasm-opt discovery was implemented twice.
- bytes_to_mb: build.rs inlined the same arithmetic build_wasi already factored.

Also fix build_wasi's EXCLUDED_LIBS comment, which falsely claimed it 'mirrors'
build.rs to 'keep the two paths in sync' — the lists deliberately differ.

No generated-code or build-flag changes; codegen drift check is clean.
